# http-server/docs/swagger.yaml
basePath: /api/v1
definitions:
  models.Info:
    properties:
      born_at:
        type: string
      dead_at:
        type: string
      id:
        type: string
      reason:
        type: string
    type: object
...
  models.Resp-models_Info:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/models.Info'
      msg:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /cow/{uuid}:
    get:
      consumes:
      - application/json
      parameters:
      - description: json
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: xiangying
          schema:
            $ref: '#/definitions/models.Resp-models_Info'
      security:
      - ApiKeyAuth: []
      summary: get uuid info
      tags:
      - getInfo
...
swagger: "2.0"